infra/requirements.txt
aws-cdk.core==1.27.0
cdk-chalice==0.6.1
aws-cdk.aws-s3==1.27.0
aws-cdk.aws-ec2==1.27.0
aws-cdk.aws-rds==1.27.0
aws-cdk.aws-events-targets==1.27.0